+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 2.0.0
+
+- Added: `--src-css-template` agument.
+- Added: `--css-template-class-name` argument.
+- Added: `--css-template-font-path` argument.
+- Added: `--css-template-font-name` argument.
+- Changed: remove `--css-template-format` argument, now format is taken from `--dest-css-template`
+- Changed: remove `--css` argument, css now generated if you use `--dest-css-template` argument.
+- Changed: rename `--css-template-dest` argument to `--dest-css-template`.
+- Remove: `--css-template` argument.
+
 # 1.0.1
 
 - Fixed: get `fontId` from `fontName`, if `fontId` is `null` or `undefined`.